feat: category details date filter
