+++ apps/web/src/app/(dashboard)/categories/[id]/page.tsx
@@ -1,9 +1,15 @@
 import { CategoryDetails } from "@/components/categories/CategoryDetails";
 import { EditCategory } from "@/components/categories/EditCategory";
 import { PageHeader } from "@/components/PageHeader";
+import { getLastMonthRange } from "@/lib/date";
 import { getQueryClient } from "@/lib/getQueryClient";
-import { getCategoryOptions } from "@/lib/queryOptions/categories";
+import {
+  getCategoryOptions,
+  getCategoryTransactionsOptions,
+} from "@/lib/queryOptions/categories";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
+
+export const dynamic = "force-dynamic";
 
 type CategoryDetailsPageProps = {
   params: Promise<{ id: string }>;
@@ -11,9 +17,15 @@ type CategoryDetailsPageProps = {
 
 const CategoryDetailsPage = async ({ params }: CategoryDetailsPageProps) => {
   const { id: categoryId } = await params;
+  const lastMonthRange = getLastMonthRange();
 
   const queryClient = getQueryClient();
-  await queryClient.prefetchQuery(getCategoryOptions(categoryId));
+  await Promise.all([
+    queryClient.prefetchQuery(getCategoryOptions(categoryId)),
+    queryClient.prefetchQuery(
+      getCategoryTransactionsOptions(categoryId, lastMonthRange),
+    ),
+  ]);
 
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
