+++ apps/web/src/components/transactions/TransactionTabs.tsx
@@ -3,6 +3,7 @@
 import { TransactionsChart } from "@/components/charts/TransactionsChart";
 import { CountUpWrapper } from "@/components/CountUpWrapper";
 import { TransactionTable } from "@/components/tables/TransactionTable";
+import type { AnalyticsDateRange } from "@/lib/date";
 import { CountType } from "@/types/dashboard";
 import type { TransactionWithRelations } from "@/types/transactions";
 import {
@@ -17,12 +18,15 @@ import {
   TabsTrigger,
 } from "@coin-guard/ui";
 import { AreaChart, Table2 } from "@coin-guard/ui/icons";
+import type { ReactNode } from "react";
 
 type TransactionTabsProps = {
   transactions: TransactionWithRelations[];
   title: string;
   description: string;
   budgetAmount?: number;
+  actions?: ReactNode;
+  range: AnalyticsDateRange;
 };
 
 export const TransactionTabs = ({
@@ -30,21 +34,26 @@ export const TransactionTabs = ({
   title,
   description,
   budgetAmount,
+  actions,
+  range,
 }: TransactionTabsProps) => {
   const sum = transactions.reduce((acc, t) => (acc += t.amount), 0);
   const average = transactions.length === 0 ? 0 : sum / transactions.length;
 
   return (
     <Tabs className="w-full" defaultValue="graph">
-      <TabsList className="min-w-64">
-        <TabsTrigger value="graph">
-          <AreaChart /> Graph
-        </TabsTrigger>
-        <TabsTrigger value="table">
-          <Table2 />
-          Table
-        </TabsTrigger>
-      </TabsList>
+      <div className="flex items-center justify-between">
+        <TabsList className="min-w-64">
+          <TabsTrigger value="graph">
+            <AreaChart /> Graph
+          </TabsTrigger>
+          <TabsTrigger value="table">
+            <Table2 />
+            Table
+          </TabsTrigger>
+        </TabsList>
+        {actions}
+      </div>
       <Card className="py-0 gap-0">
         <CardHeader className="flex flex-col items-stretch border-b p-0! sm:flex-row">
           <div className="flex flex-1 flex-col justify-center gap-1 px-6 pt-4 pb-3 sm:py-0!">
@@ -70,6 +79,7 @@ export const TransactionTabs = ({
           <TabsContent value="graph">
             <TransactionsChart
               budgetAmount={budgetAmount}
+              range={range}
               transactions={transactions}
             />
           </TabsContent>
