+++ apps/web/src/lib/queryOptions/categories.ts
@@ -1,6 +1,8 @@
 import { GetCategories } from "@/actions/categories/GetCategories";
 import { GetCategory } from "@/actions/categories/GetCategory";
+import { GetCategoryTransactions } from "@/actions/categories/GetCategoryTransactions";
 import { KEYS } from "@/constants/queryKeys";
+import type { AnalyticsDateRange } from "@/lib/date";
 import { queryOptions } from "@tanstack/react-query";
 
 export const getCategoriesOptions = () => {
@@ -14,5 +16,15 @@ export const getCategoryOptions = (categoryId: string) => {
   return queryOptions({
     queryKey: KEYS.category(categoryId),
     queryFn: () => GetCategory(categoryId),
+  });
+};
+
+export const getCategoryTransactionsOptions = (
+  categoryId: string,
+  range: AnalyticsDateRange,
+) => {
+  return queryOptions({
+    queryKey: KEYS.categoryTransactions(categoryId, range.from, range.to),
+    queryFn: () => GetCategoryTransactions(categoryId, range),
   });
 };
