+++ apps/web/src/store/categoryStore.ts
@@ -1,0 +1,7 @@
+import type { AnalyticsDateRange } from "@/lib/date";
+import { getLastMonthRange } from "@/lib/date";
+import { atom } from "jotai";
+
+export const categoryTransactionsRangeAtom = atom<AnalyticsDateRange>(
+  getLastMonthRange(),
+);
