+++ apps/web/src/types/categories.ts
@@ -1,9 +1,5 @@
 import type { Prisma } from "@coin-guard/db";
 
-export type CategoryWithTransactions = Prisma.CategoryGetPayload<{
-  include: { transactions: true };
-}>;
-
 export type CategoryWithLookups = Prisma.CategoryGetPayload<{
   include: { lookups: true };
 }>;
